Anhlh

+++ Areas/Users/Controllers/UsersController.cs
@@ -152,9 +152,13 @@ namespace MAQFurni.Areas_Users_Controllers
         public async Task<IActionResult> DeleteConfirmed(string id)
         {
             var user = await _context.Users.FindAsync(id);
-            _context.Users.Remove(user);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            var order = _context.Orders.Where(o => o.UserId == id);
+            if (order == null)
+            {
+                _context.Users.Remove(user);
+                await _context.SaveChangesAsync();
+            }
+            return RedirectToAction(nameof(Index), new { status = "fail" });
         }
 
         private bool UserExists(string id)
